iw2 is an integer array not real
Former-commit-id: ce11362e137cb5294f00497b5aba347194222ff6

+++ IM/CRCM/src/ModFieldTrace.f90
@@ -10,7 +10,7 @@ Module ModFieldTrace
 
   integer :: irm(ip),irm0(ip),iba(ip)
 
-  real :: iw2(ik)
+  integer :: iw2(ik)
 
   logical :: UseEllipse = .true.
 contains
